django-postgres/requirements.txt
# ── Web Framework ──────────────────────────────────────────
Django==5.0.3

# ── REST API ────────────────────────────────────────────────
djangorestframework==3.15.1

# ── PostgreSQL Driver ───────────────────────────────────────
psycopg2-binary==2.9.9

# ── Environment Variable Management ────────────────────────
python-decouple==3.8

# ── Production WSGI Server ──────────────────────────────────
gunicorn==21.2.0
